requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-experimental
langchain-google-genai
pandas
matplotlib
seaborn
python-dotenv
requests
pydantic
